Work through support lib clang-tidy warnings.
Move configuration from clang_tidy.cmake to .clang-tidy. Explicitly
request the correct casing style from clang-tidy for all
categories. Add macro configuration (UPPER_CASE and PSTORE_ prefix).

+++ lib/support/uint128.cpp
@@ -42,7 +42,8 @@ namespace {
             if (!digit) {
                 return nothing<std::uint64_t> ();
             }
-            result |= (static_cast<std::uint64_t> (digit.value ()) << shift);
+            result |=
+                (static_cast<std::uint64_t> (digit.value ()) << static_cast<unsigned> (shift));
         }
         return just (result);
     }
